requirements.txt
google-api-python-client==2.108.0
pandas==2.1.4
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==1.4.51
requests==2.31.0
python-dotenv==1.0.0